+++ naas_drivers/tools/googleanalytics.py
@@ -1,12 +1,10 @@
 """Google Analytics Driver."""
 import re
 from datetime import datetime, timedelta
-
 import numpy as np
 import pandas as pd
 from google.oauth2 import service_account
 from apiclient.discovery import build
-
 from naas_drivers.driver import InDriver, OutDriver
 
 
@@ -17,6 +15,50 @@ def ga_naming_to_title(ga_nanimg: str):
     return " ".join([name.title() for name in splited_name])
 
 
+ga_metrics = [
+    "ga:users",
+    "ga:newUsers",
+    "ga:percentNewSessions",
+    "ga:sessions",
+    "ga:bounces",
+    "ga:bounceRate",
+    "ga:sessionDuration",
+    "ga:avgSessionDuration",
+    "ga:organicSearches",
+    "ga:entrances",
+    "ga:entranceRate",
+    "ga:pageviews",
+    "ga:pageviewsPerSession",
+    "ga:uniquePageviews",
+    "ga:timeOnPage",
+    "ga:avgTimeOnPage",
+    "ga:exits",
+    "ga:exitRate",
+    "ga:pageLoadTime",
+    "ga:pageLoadSample",
+    "ga:avgPageLoadTime",
+    "ga:domainLookupTime",
+    "ga:avgDomainLookupTime",
+    "ga:pageDownloadTime",
+    "ga:avgPageDownloadTime",
+    "ga:redirectionTime",
+    "ga:avgRedirectionTime",
+    "ga:serverConnectionTime",
+    "ga:avgServerConnectionTime",
+    "ga:serverResponseTime",
+    "ga:avgServerResponseTime",
+    "ga:speedMetricsSample",
+    "ga:domInteractiveTime",
+    "ga:avgDomInteractiveTime",
+    "ga:domContentLoadedTime",
+    "ga:avgDomContentLoadedTime",
+    "ga:domLatencyMetricsSample",
+    "ga:uniqueDimensionCombinations",
+    "ga:hits",
+    "ga:sessionsPerUser",
+]
+
+
 class GoogleAnalytics(InDriver, OutDriver):
     """
     Google Analytics driver.
@@ -24,13 +66,145 @@ class GoogleAnalytics(InDriver, OutDriver):
 
     def __init__(self) -> None:
         self.views = Views(self)
+        self.available_metrics = []
+        self.__generate_methods()
 
-    def connect(self, json_path: str):
+    def connect(self, json_path: str, view_id: str):
         credentials = service_account.Credentials.from_service_account_file(
             json_path, scopes=["https://www.googleapis.com/auth/analytics.readonly"]
         )
+        self.view_id = view_id
         self.service = build("analyticsreporting", "v4", credentials=credentials)
         return self
+
+    # This method is used to automatically generate methods based on metrics available in GA.
+    def __generate_methods(self):
+        for metric_id in ga_metrics:
+
+            computed_name = metric_id
+            computed_name = re.sub(r"(?<!^)(?=[A-Z])", "_", computed_name).lower()[3:]
+
+            try:
+                getattr(self, computed_name)
+            except Exception as e:  # noqa: F841
+                self.available_metrics.append(computed_name)
+                setattr(self, computed_name, Metric())
+
+            def custom_get_trend(metric_id):
+                return (
+                    lambda dimensions, start_date=None, end_date=None: self.get_trend(
+                        metric_id, dimensions, start_date, end_date
+                    )
+                )
+
+            setattr(
+                getattr(self, computed_name), "get_trend", custom_get_trend(metric_id)
+            )
+
+    def get_trend(self, metrics, dimensions, start_date, end_date):
+        """
+        Return an dataframe object with 6 columns:
+        - DATE         GA dimensions
+        - METRIC       GA metrics
+        - VALUE        Metrics value
+        - VALUE_COMP   Metrics last value comparison
+        - VARV         Variation in value between VALUE and VALUE_COMP
+        - VARP         Variation in % between VALUE and VALUE_COMP
+
+        Parameters
+        ----------
+        metrics: str:
+            - New visitors = "ga:newUsers"
+            - User = "ga:users"
+        dimensions: str:
+            List of google analytics dimensions
+            - Hourly = "hourly"
+            - Day = "daily"
+            - Week = "weekly"
+            - Month = "monthly"
+        start_date: str: default="30daysAgo"
+            "NdaysAgo" with n equal to int
+        end_date: str: default="today"
+            "today", "yesterday"
+        """
+
+        allowed_aliases = ["hourly", "daily", "weekly", "monthly"]
+        allowed_dimensions = [
+            "ga:date,ga:hour",
+            "ga:date",
+            "ga:year,ga:week",
+            "ga:year,ga:month",
+        ]
+
+        if dimensions in allowed_aliases:
+            dimensions = allowed_dimensions[allowed_aliases.index(dimensions)]
+
+        if dimensions not in allowed_dimensions:
+            raise Exception(f"'dimensions' should be one of {allowed_dimensions}")
+
+        # Get data
+        df = self.views.get_data(
+            self.view_id,
+            metrics=metrics,
+            dimensions=dimensions,
+            start_date=start_date,
+            end_date=end_date,
+            format_type="summary",
+            pivots_dimensions="ga:country",  # not used
+        )
+
+        # Format trend dataset
+        df["DATE_ISO"] = df.index
+        df = df.reset_index(drop=True)
+        if dimensions == "ga:date,ga:hour":
+            df["DATE_ISO"] = pd.to_datetime(
+                df.apply(lambda row: f"{row.DATE_ISO[0]} {row.DATE_ISO[1]}:00", axis=1)
+            )
+            df["DATE"] = df["DATE_ISO"].dt.strftime("%Y-%m-%d %H:00:00")
+        elif dimensions == "ga:date":
+            df["DATE_ISO"] = pd.to_datetime(
+                df.apply(lambda row: row.DATE_ISO[0], axis=1)
+            )
+            df["DATE"] = df["DATE_ISO"].dt.strftime("%Y-%m-%d")
+        elif dimensions == "ga:year,ga:week":
+            df["DATE_ISO"] = pd.to_datetime(
+                df.apply(
+                    lambda row: datetime.strptime(
+                        f"{row.DATE_ISO[0]}-W{row.DATE_ISO[1]}" + "-1", "%Y-W%W-%w"
+                    ),
+                    axis=1,
+                )
+            )
+            df["DATE"] = df["DATE_ISO"].dt.strftime("%Y W%W")
+        elif dimensions == "ga:year,ga:month":
+            df["DATE_ISO"] = pd.to_datetime(
+                df.apply(
+                    lambda row: datetime.strptime(
+                        f"{row.DATE_ISO[0]}-M{row.DATE_ISO[1]}", "%Y-m%m"
+                    ),
+                    axis=1,
+                )
+            )
+            df["DATE"] = df["DATE_ISO"].dt.strftime("%Y %b")
+        df["METRIC"] = metrics.replace("ga:", "")
+        df["VALUE"] = df[metrics]
+        df = df.drop(metrics, axis=1)
+        df.columns = df.columns.str.upper()
+
+        # Calc variation
+        for idx, row in df.iterrows():
+            if idx == 0:
+                value_n1 = 0
+            else:
+                value_n1 = df.loc[df.index[idx - 1], "VALUE"]
+            df.loc[df.index[idx], "VALUE_COMP"] = value_n1
+        df["VARV"] = df["VALUE"] - df["VALUE_COMP"]
+        df["VARP"] = df["VARV"] / abs(df["VALUE_COMP"])
+        return df
+
+
+class Metric:
+    pass
 
 
 class Views:
@@ -45,6 +219,7 @@ class Views:
         metrics: str,
         pivots_dimensions: str,
         dimensions: str = "ga:yearMonth",
+        max_group_count: int = 1000,
     ) -> dict:
         """
         Create the body of the request to Google Analytics Reporting API V4.
@@ -63,12 +238,17 @@ class Views:
                 {
                     "viewId": view_id,
                     "dateRanges": {"startDate": start_date, "endDate": end_date},
-                    "metrics": [{"expression": metrics}],
-                    "dimensions": {"name": dimensions},
+                    "metrics": [
+                        {"expression": metric} for metric in metrics.split(",")
+                    ],
+                    "dimensions": [
+                        {"name": dimension} for dimension in dimensions.split(",")
+                    ],
                     "pivots": [
                         {
                             "dimensions": {"name": pivots_dimensions},
                             "metrics": [{"expression": metrics}],
+                            "maxGroupCount": max_group_count,
                         }
                     ],
                 }
@@ -84,6 +264,7 @@ class Views:
         start_date: str = None,
         end_date: str = None,
         format_type: str = "summary",
+        max_group_count: int = 1000,
     ) -> pd.DataFrame:
         """
         Get data from Google Analytics Reporting API V4.
@@ -101,7 +282,13 @@ class Views:
         end_date = end_date if end_date else datetime.today().strftime("%Y-%m-%d")
         # Create body
         body = self._get_body(
-            view_id, start_date, end_date, metrics, pivots_dimensions, dimensions
+            view_id,
+            start_date,
+            end_date,
+            metrics,
+            pivots_dimensions,
+            dimensions,
+            max_group_count=max_group_count,
         )
         # Fetch Data
         try: